Scrubbing mscorlib.cs and move COM interop types into interop.cs
